fix: Adds Poison to codegen
