+++ .config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "modernuoschemagenerator": {
-      "version": "2.0.3",
+      "version": "2.0.4",
       "commands": [
         "ModernUOSchemaGenerator"
       ]
