--- Projects/Server/Serialization/IRawSerializable.cs
@@ -1,23 +1,0 @@
-/*************************************************************************
- * ModernUO                                                              *
- * Copyright 2019-2021 - ModernUO Development Team                       *
- * Email: hi@modernuo.com                                                *
- * File: IRawSerializable.cs                                             *
- *                                                                       *
- * This program is free software: you can redistribute it and/or modify  *
- * it under the terms of the GNU General Public License as published by  *
- * the Free Software Foundation, either version 3 of the License, or     *
- * (at your option) any later version.                                   *
- *                                                                       *
- * You should have received a copy of the GNU General Public License     *
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
- *************************************************************************/
-
-namespace Server
-{
-    public interface IRawSerializable
-    {
-        void Deserialize(IGenericReader reader);
-        void Serialize(IGenericWriter writer);
-    }
-}
