+++ Projects/Server/Server.csproj
@@ -39,7 +39,7 @@
         <PackageReference Include="Zlib.Bindings" Version="1.8.0" />
 
         <PackageReference Include="ModernUO.Serialization.Annotations" Version="2.0.2" />
-        <PackageReference Include="ModernUO.Serialization.Generator" Version="2.0.3" />
+        <PackageReference Include="ModernUO.Serialization.Generator" Version="2.0.4" />
     </ItemGroup>
     <ItemGroup>
         <AdditionalFiles Include="Migrations/*.v*.json" />
