+++ Projects/UOContent/Items/Skill Items/Ninjitsu/Shuriken.cs
@@ -94,7 +94,7 @@ namespace Server.Items
 
             writer.Write(m_UsesRemaining);
 
-            Poison.Serialize(m_Poison, writer);
+            writer.Write(m_Poison);
             writer.Write(m_PoisonCharges);
         }
 
@@ -110,7 +110,7 @@ namespace Server.Items
                     {
                         m_UsesRemaining = reader.ReadInt();
 
-                        m_Poison = Poison.Deserialize(reader);
+                        m_Poison = reader.ReadPoison();
                         m_PoisonCharges = reader.ReadInt();
 
                         break;
