+++ Projects/UOContent/UOContent.csproj
@@ -46,7 +46,7 @@
         <PackageReference Include="Zstd.Binaries" Version="1.3.0" />
 
         <PackageReference Include="ModernUO.Serialization.Annotations" Version="2.0.2" />
-        <PackageReference Include="ModernUO.Serialization.Generator" Version="2.0.3" />
+        <PackageReference Include="ModernUO.Serialization.Generator" Version="2.0.4" />
     </ItemGroup>
     <ItemGroup>
         <AdditionalFiles Include="Migrations/*.v*.json" />
